Move GA code from _affiliation_network.py to _evolve.py
